+++ obs-studio-client/source/volmeter.cpp
@@ -51,7 +51,7 @@ osn::Volmeter::Volmeter(const Napi::CallbackInfo &info) : Napi::ObjectWrap<osn::
 {
 	Napi::Env env = info.Env();
 	Napi::HandleScope scope(env);
-	int length = info.Length();
+	size_t length = info.Length();
 
 	if (length <= 0 || !info[0].IsNumber()) {
 		Napi::TypeError::New(env, "Number expected").ThrowAsJavaScriptException();
@@ -77,8 +77,8 @@ Napi::Value osn::Volmeter::Create(const Napi::CallbackInfo &info)
 	if (!ValidateResponse(info, response))
 		return info.Env().Undefined();
 
-	auto instance = osn::Volmeter::constructor.New(
-		{Napi::Number::New(info.Env(), response[1].value_union.ui64), Napi::Number::New(info.Env(), response[2].value_union.ui32)});
+	auto instance = osn::Volmeter::constructor.New({Napi::Number::New(info.Env(), static_cast<double>(response[1].value_union.ui64)),
+							Napi::Number::New(info.Env(), static_cast<double>(response[2].value_union.ui32))});
 
 	globalCallback::add_volmeter(response[1].value_union.ui64);
